+++ Payloads/BanPlayerPayload.cs
@@ -1,0 +1,9 @@
+
+namespace WebAPI.Payloads
+{
+    public class BanPlayerPayload
+    {
+        public string reason { get; set; }
+        public double hours { get; set; }
+    }
+}