+++ Payloads/KickPlayerPayload.cs
@@ -1,0 +1,8 @@
+
+namespace WebAPI.Payloads
+{
+    public class KickPlayerPayload
+    {
+        public string reason { get; set; }
+    }
+}